Onyx Boox T68 Lynx: Full support for screen dithering. Workaround for the spurious A2 mode switch problem in older firmwares.
Uses switching to dithering,inverted mode; which is reliable. As a
consequence, the screen has to be inverted, implemented in TopCanvas
just before buffer flipping. A2 mode does not need to be activated
manually.

This is not necessary on newer firmwares (late 2015 and newer), where A2
switching is reliable. A2 then needs to be switched on manually however.

Conflicts:
	src/Android/Main.cpp