Map Elements / Aircraft: Handle line widths consistently with other line widths in the system
This removes the following special case: In the case of OpenGL
rendering, the line width for the aircraft symbol was handled
differently from other line widths: it was in a logical line width unit
rather than in pixel units. As now all line widths in OpenGL rendering
are in logical units to be pixel density independent, the special
handling of this case is removed. Except for a different scale factor
the OpenGL code is now identical to the non-GL code.

As a result, the symbol coordinates are no longer rotated and scaled via
OpenGL, but rather in the CPU. However, due to the low number of points
in the symbol this hardly is a performance loss.